Implement headerBuilder class
For now a basic version has been implemented but this class will later on be able to provide a lot more flexibility in header information for each page.

Currently it just creates a header from the given parameters (that is a title), but future plans could be different stylesheets per page and more general versitility.